+++ udata_front/tests/views/test_site.py
@@ -270,7 +270,7 @@ class SiteViewsTest(GouvfrFrontTestCase):
     def test_organizations_csv(self):
         self.app.config['EXPORT_CSV_MODELS'] = []
         orgs = [OrganizationFactory() for _ in range(5)]
-        hidden_org = OrganizationFactory(deleted=datetime.now())
+        hidden_org = OrganizationFactory(deleted=datetime.utcnow())
 
         response = self.get(url_for('site.organizations_csv'))
 
@@ -408,7 +408,7 @@ class SiteViewsTest(GouvfrFrontTestCase):
                   for i in range(5)]
         hidden_harvest = HarvestSource.objects.create(
             url='https://example.com/deleted',
-            deleted=datetime.now()
+            deleted=datetime.utcnow()
         )
 
         response = self.get(url_for('site.harvests_csv'))